Set `composite: false` by default to avoid .tsbuildinfo files in package archives (#2)
Installing a package based on this tsconfig from a git commit could lead to checksum mismatches, because the `.tsbuildinfo` was included in the package archive but its contents are platform-specific (for example, filenames are lowercased on macOS to support case-insensitivity).

By setting `composite: false`, we don't produce any .tsbuildinfo files by default. Individual projects may enable this setting if desired, though they should take care not to include the .tsbuildinfo in build products.